Backend/requirements.txt
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
black==24.1.1
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.9
django-allauth==0.61.1
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
elastic-transport==8.13.0
elasticsearch==8.13.0
elasticsearch-dsl==8.12.0
exceptiongroup==1.2.0
flake8==7.0.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaggle==1.5.12
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==6.1.1
python3-openid==3.2.0
pytz==2023.4
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
